bugfix Vagrantfile shell provisioner on 1st and 2nd boot

+++ Vagrantfile
@@ -5,10 +5,15 @@ VAGRANTFILE_API_VERSION = "2"
 
 pubKeyFile = 'id_rsa.pub'
 pubKeyPath = ENV['HOME'] + '/.ssh/' + pubKeyFile
-if File.exist?( pubKeyPath )
-	baseMachine = '.vagrant/machines/base/virtualbox'
+baseMachine = '.vagrant/machines/base/virtualbox'
+pubKeyForRoot = 'echo "PLEASE RUN [ vagrant reload --provision ] AFTER FIRST BOOT"'
+if File.exist?( pubKeyPath ) && File.exist?( baseMachine )
 	FileUtils.cp( pubKeyPath, baseMachine )
-	pubKeyForRoot = 'cat /vagrant/' + baseMachine + '/' + pubKeyFile + ' > /root/.ssh/authorized_keys'
+	sshDir = '/root/.ssh'
+	sshKeys = sshDir + '/authorized_keys'
+	pubKeyForRoot = 'mkdir -p ' + sshDir +
+	 	' && cat /vagrant/' + baseMachine + '/' + pubKeyFile + ' > ' + sshKeys +
+		' && echo "YOU MAY NOW [ ssh root@local ] WITHOUT A PASSWORD"'
 end
 
 Vagrant.configure(VAGRANTFILE_API_VERSION) do |config|
@@ -19,7 +24,6 @@ Vagrant.configure(VAGRANTFILE_API_VERSION) do |config|
 		base.vm.synced_folder ENV['HOME'] + "/sync", "/sync", create: true, type: "nfs"
 		base.vm.network "private_network", ip: "192.168.200.2"
 		base.vm.network "forwarded_port", guest: 80, host: 8080
-		base.vm.network "forwarded_port", guest: 443, host: 8443
 		base.vm.boot_timeout = 300
 		base.vm.provider :virtualbox do |vb|
 			vb.name = "base"
@@ -28,3 +32,4 @@ Vagrant.configure(VAGRANTFILE_API_VERSION) do |config|
 	end
 
 end
+
